refactor: rewrite from JavaScript to TypeScript
